About Page Main Content
Added JS to make buttons that reveal the main content for my about page. Also added a cool button for those transitions. Right now it it uses a loop that goes through the button under the class name "wrapperExpButton". I still need to add last button as well as animations when the content is loaded in.